+++ text-runner/src/actions/external-action-manager.ts
@@ -1,7 +1,7 @@
-import { FunctionRepo } from "./action-finder"
 import { Action } from "./types/action"
-import { UnprintedUserError } from "../errors/unprinted-user-error"
 import { actionName } from "./helpers/action-name"
+import { FunctionRepo } from "./types/function-repo"
+import { UnprintedUserError } from "../errors/unprinted-user-error"
 
 /** ExternalActionManager provides external actions from their own NPM modules */
 export class ExternalActionManager {
