+++ text-runner/src/actions/types/function-repo.ts
@@ -1,0 +1,5 @@
+import { Action } from "./action"
+
+export interface FunctionRepo {
+  [key: string]: Action
+}
